requirements.txt
fastapi==0.120.3
uvicorn==0.38.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
pydantic==2.12.3
google-generativeai==0.7.1
pandas==2.3.3
numpy==2.3.4
matplotlib==3.10.7
streamlit==1.42.0
requests==2.32.3